Renaming Main Bot
